requirements.txt
# Python dependencies
flask==3.0.2
flask-cors==4.0.0
langchain-core==0.1.27
langchain-mistralai==0.0.11
python-dotenv==1.0.1

# JavaScript dependencies
{
  "dependencies": {
    "axios": "^1.6.7",
    "cors": "^2.8.5"
  }
}

